pyproject.toml
[tool.poetry]
name = "projeto"
version = "0.1.0"
description = ""
authors = ["Lucas-Sobreira <lucas-andrade2011@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
requests = "^2.32.3"
apache-airflow = "^2.10.2"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"  